requirements.txt
PyYAML>=6.0
aiohttp>=3.9
aiohttp-socks>=0.8
rich>=13.0
